// server/src/auth/auth.enum.ts
export enum UserStatus {
  NORMAL = 'NORMAL',
  FORBIDDEN = 'FORBIDDEN',
  DELETED = 'DELETED',
}
export enum AdminStatus {
  NORMAL = 'NORMAL',
  FORBIDDEN = 'FORBIDDEN',
  DELETED = 'DELETED',
}

export enum UserRole {
  NORMAL = 'NORMAL',
  EDITOR = 'EDITOR',
  ADMIN = 'ADMIN',
  SUPERADMIN = 'SUPERADMIN',
}